implemented heapifyDown for delete and build Heap
heapifyDown is necessary for deletion(yet to be implemented) and
building an heap from a given array, remember that the array is supposed
to be 1 index array, in which the leaf nodes starts from floor(size/2) + 1.

that means if it was a 0 index array then the leaf nodes would have been
started from floor(size/2).